myapp/requirements.txt
Flask==2.0.1
openai==0.10.2
python-dotenv==0.19.0
PyPDF2==1.26.0
faiss-cpu==1.7.1.post2
python-docx
pandas==1.3.3
python-docx==0.8.10
openpyxl==3.0.7
langchain==0.1.0
